requirements.txt
ffmpeg-python==0.2.0
pydub==0.25.1
SpeechRecognition==3.10.0
numpy==1.24.3
scipy==1.10.1
librosa==0.10.0
torch==2.0.1
torchaudio==2.0.2
transformers==4.30.2
sentencepiece==0.1.99
pyaudio==0.2.13
matplotlib==3.7.2
pysrt==1.1.2
whisper==1.1.10
vosk==0.3.45
tqdm==4.65.0
gradio==3.50.2